+++ lib/clixon/clixon_netns.h
@@ -1,0 +1,14 @@
+/*
+ * Network namespace code 
+ * @thanks Anders Franzén, especially get_sock() and send_sock() functions
+*/
+
+#ifndef _CLIXON_NETNS_H_
+#define _CLIXON_NETNS_H_
+
+/*
+ * Prototypes
+ */
+int clixon_netns_socket(const char *netns, struct sockaddr *sa, size_t sin_len, int backlog, int *sock);
+
+#endif  /* _CLIXON_NETNS_H_ */